requirements.txt
streamlit
langchain
langchain-openai
langchain-community
python-dotenv
tiktoken
faiss-cpu
sentence-transformers
pypdf
duckduckgo-search
langchainhub
google-search-results
langchain_experimental
matplotlib
tabulate
openpyxl
mcp
